Separate spotify auth and api components
